requirements.txt
anyio==4.4.0
APScheduler==3.10.4
cachetools==5.5.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
python-telegram-bot==21.5
pytz==2024.2
PyYAML==6.0.2
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
